crud para documentos

+++ FlashMindAPI/src/main/java/br/com/roselabs/flashmind/repositories/DocumentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentRepository extends CrudRepository<Document, Integer> {
+public interface DocumentRepository extends CrudRepository<Document, Long> {
 
 }
